requirements.txt
# Web & Backend
Flask==3.0.3
gunicorn==21.2.0
markupsafe

# RSS Parsing
feedparser==6.0.11

# OpenAI API
openai
python-dotenv

# Data Handling
pandas>=2.0
numpy>=1.24.0

# Visualization
altair==5.3.0
vega-datasets==0.9.0
matplotlib>=3.7.0
seaborn>=0.12.2

# Date & Time Utilities
python-dateutil>=2.8.2

#API requests and parsing
requests>=2.31.0
beautifulsoup4>=4.12.0

